CV-10 make getter also return short

+++ src/main/java/be/dominionexperts/cvtool/dto/Keyword.java
@@ -28,7 +28,7 @@ public class Keyword {
         return name;
     }
 
-    public int getLevel() {
+    public short getLevel() {
         return level;
     }
